Add MediaItemModal.tsx
